+++ src/presentation/protocools/index.ts
@@ -1,0 +1,3 @@
+export * from "./controller";
+export * from "./emailValidator";
+export * from "./http";
